Minor robustness fixes: overflow-safe binary search midpoint and prime-check loop, keep ll weights in Prim, reset bipartite counts with 0, guard multi-ordered-set erase and document its two-part paste

+++ code/graphs/prim.cpp
@@ -7,7 +7,7 @@ function<ll()> prim = [&]() -> ll
   {
     pll p = pq.top();
     pq.pop();
-    int w = p.fi;
+    ll w = p.fi;
     int u = p.se;
     if( visited[u] ) continue;
     res += w;
